requirements.txt
googlesearch-python
google-generativeai
requests
python-dotenv
beautifulsoup4
